backend/requirements.txt
django>=5.0,<6.1
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
django-allauth>=65.0
dj-rest-auth[with_social]>=7.0
psycopg2-binary>=2.9
Pillow>=10.0
openpyxl>=3.1
reportlab>=4.0
python-decouple>=3.8
google-auth>=2.0
firebase-admin>=6.0
